+++ src/Strategy/BerkeleyDB.hs
@@ -4,7 +4,7 @@ module Strategy.BerkeleyDB (
   mkProject,
 ) where
 
-import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProject), analyzeProject')
+import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProject), analyzeProjectStaticOnly)
 import Container.OsRelease (OsInfo (..))
 import Control.Effect.Diagnostics (Diagnostics, context)
 import Control.Effect.Lift (Lift)
@@ -48,7 +48,7 @@ instance ToJSON BerkeleyDatabase
 
 instance AnalyzeProject BerkeleyDatabase where
   analyzeProject _ = getDeps
-  analyzeProject' _ = getDeps
+  analyzeProjectStaticOnly _ = getDeps
 
 discover ::
   ( Has ReadFS sig m