+++ src/Strategy/Gradle.hs
@@ -18,7 +18,7 @@ module Strategy.Gradle (
   GradleProject,
 ) where
 
-import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProject'), analyzeProject)
+import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProjectStaticOnly), analyzeProject)
 import App.Fossa.Config.Analyze (ExperimentalAnalyzeConfig (allowedGradleConfigs))
 import Control.Algebra (Has)
 import Control.Effect.Diagnostics (
@@ -169,7 +169,7 @@ instance ToJSON GradleProject
 
 instance AnalyzeProject GradleProject where
   analyzeProject = getDeps
-  analyzeProject' _ = const $ fatalText "Cannot analyze Gradle target statically"
+  analyzeProjectStaticOnly _ = const $ fatalText "Cannot analyze Gradle target statically"
 
 gradleProjectsCmd :: Text -> Command
 gradleProjectsCmd baseCmd =