+++ src/Strategy/NuGet/ProjectAssetsJson.hs
@@ -13,7 +13,7 @@ module Strategy.NuGet.ProjectAssetsJson (
   FrameworkName (..),
 ) where
 
-import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProject'), analyzeProject)
+import App.Fossa.Analyze.Types (AnalyzeProject (analyzeProjectStaticOnly), analyzeProject)
 import Control.Effect.Diagnostics (
   Diagnostics,
   Has,
@@ -91,7 +91,7 @@ instance ToJSON ProjectAssetsJsonProject
 
 instance AnalyzeProject ProjectAssetsJsonProject where
   analyzeProject _ = getDeps
-  analyzeProject' _ = getDeps
+  analyzeProjectStaticOnly _ = getDeps
 
 mkProject :: ProjectAssetsJsonProject -> DiscoveredProject ProjectAssetsJsonProject
 mkProject project =